airflow_project/requirements.txt
pandas
scikit-learn
xgboost
mlflow>=2.0.0
alibi-detect
databricks-sql-connector
pyyaml
dotenv
seaborn